requirements.txt
django-q==1.3.9
Django==4.2
djangorestframework==3.14
django-cors-headers
drf-spectacular-sidecar==2024.7.1
drf-spectacular==0.27.2
ffmpeg-python==0.2.0
openai-whisper
opencv-python==4.7.0.72
opensearch-py==2.3.1
opentelemetry-api==1.26.0
opentelemetry-sdk==1.26.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
scikit-image==0.20.0
transformers==4.30.2
yt-dlp>=2024.04.09
Pillow==10.1.0
torch==2.1.0
torchvision==0.16.0
numpy==1.26.1
open-clip-torch==2.20.0
sentence-transformers==2.2.2
